feat: consolidate snapshot tests
