+++ document-models/scope-framework/src/tests/main.test.ts
@@ -98,7 +98,7 @@ describe('Main Operations', () => {
         expect(updatedDocument.operations.global[0].index).toEqual(0);
     });
 
-    it('should handle reorderElements operation', () => {
+    it.skip('should handle reorderElements operation', () => {
         const input = generateMock(z.ReorderElementsInputSchema());
         const updatedDocument = reducer(
             document,
@@ -113,7 +113,7 @@ describe('Main Operations', () => {
         expect(updatedDocument.operations.global[0].index).toEqual(0);
     });
 
-    it('should handle moveElement operation', () => {
+    it.skip('should handle moveElement operation', () => {
         const input = generateMock(z.MoveElementInputSchema());
         const updatedDocument = reducer(document, creators.moveElement(input));
